Implemented Greedysearch with unit testing.

+++ V2.0/header_files/medoid.h
@@ -7,4 +7,4 @@ using namespace std;
 
 
 void pfinder(vector <int> &Rf,map<int,float>&T,int &p);
-map <int,float> FindMedoid(vector<vector<float>>DataNodes,int r,set <float>& category_attributes);
+map <float,int> FindMedoid(vector<vector<float>>DataNodes,int r,set <float>& category_attributes);
